+++ yandex/billing_service_instance_binding_test.go
@@ -1,0 +1,15 @@
+package yandex
+
+import (
+	"os"
+	"time"
+)
+
+func billingInstanceTestFirstBillingAccountId() string {
+	return os.Getenv("YC_BILLING_TEST_ACCOUNT_ID_1")
+}
+func billingInstanceTestSecondBillingAccountId() string {
+	return os.Getenv("YC_BILLING_TEST_ACCOUNT_ID_2")
+}
+
+const yandexBillingServiceInstanceBindingDefaultTimeout = 1 * time.Minute